% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{binary_class}
\alias{binary_class}
\title{Performing binary classification}
\usage{
binary_class(data = NULL, class.file = NULL, scale = FALSE)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{class.file}{(Required) Path to classification file in comma-separated format. Please see tutorial for details.}

\item{scale}{Standardisation of data (Default: FALSE)}
}
\description{
This function performs binary classification on loaded data using a user-specified list of criteria. Returns a column in data frame class.
}
\examples{
binary.results <- binary_class(data, class.file = 'cell_types.csv')
}
