% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{binary_summary}
\alias{binary_summary}
\title{Summarising binary classification results}
\usage{
binary_summary(data = NULL, binary.results = NULL)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{binary.results}{(Required) Cell type results in data frame format. Expects results from binary_class}
}
\description{
This function summarises binary classification results from binary_class. Returns a data frame of median expression values of all markers, abundance and percentage abundance for all user-defined cell types
}
\examples{
binary.summary <- binary_summary(data, binary.results)
}
