% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expressions.R
\name{calculate_diff}
\alias{calculate_diff}
\title{Performing differential analysis on multiple datasets}
\usage{
calculate_diff(
  multi.data = NULL,
  column = "Cell.Type",
  remove.unclass = FALSE,
  transform = "default"
)
}
\arguments{
\item{multi.data}{(Required) A data frame of clustering results. Expects results from load_multi_samples}

\item{column}{Choose between plotting Cell.Type or Cell.Subtype (Default: Cell.Type)}

\item{remove.unclass}{Option to remove unclassified cells from plot (Default: FALSE)}

\item{transform}{Transformation method used for original analysis. 'default' assumes arcsinh transformation. 'none' indicates no transformation (Default: default)}
}
\description{
This function performs differential analysis on datasets as imported by load_multiple_samples. Returns a list of three data frames: [\link{1}] percentage abundances of each cell (sub)type in each sample, [\link{2}] log2 fold changes of each cell (sub)type between two conditions and [\link{3}] p-values for the fold changes
}
\examples{
diff.results <- calculate_diff(multi.data)
diff.results <- calculate_diff(multi.data, transform = 'default', column = 'Cell.Subtype')
}
