% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_subsets}
\alias{cluster_subsets}
\title{Clustering binary classified results}
\usage{
cluster_subsets(
  data = NULL,
  binary.results = NULL,
  method = "kmeans",
  no_of_clusters = 40
)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{binary.results}{(Required) Cell type results in data frame format. Expects results from binary_class}

\item{method}{Clustering method. k-means (Default) or FlowSOM}

\item{no_of_clusters}{Default: 40}
}
\description{
This function clusters cells individually for each user-definied cell type. Returns a data frame.
}
\examples{
cluster.results <- cluster_subsets(data, binary.results)
cluster.results <- cluster_subsets(data, binary.results, method = 'FlowSOM')
}
