% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_summary}
\alias{cluster_summary}
\title{Summarising clustering results}
\usage{
cluster_summary(data = NULL, cluster.results = NULL)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{cluster.results}{(Required) Cell type results in data frame format. Expects results from cluster_subsets}
}
\description{
This function summarises post-binary classification clustering results. Returns a data frame of median expressions and abundance for each cluster of each user-defined cell type.
}
\examples{
cluster.summary <- cluster_summary(data, cluster.results)
}
