% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{examine_data}
\alias{examine_data}
\title{Examining an FCS File}
\usage{
examine_data(input.file = NULL, transform = "arcsinh", cofactor = 5)
}
\arguments{
\item{input.file}{(Required) Path to input file}

\item{transform}{'arcsinh' or 'none'. 'arcsinh' performs an arcsinh transformation. 'none' performs no data trasformation}

\item{cofactor}{Cofactor for arcsinh transformation (Default: 5)}
}
\description{
This function reads an FCS file and plots density plots for individual markers.
}
\examples{
examine_data('example.fcs')
}
