% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_data}
\alias{load_data}
\title{Loading an FCS File}
\usage{
load_data(
  input.file = NULL,
  channels = NULL,
  transform = "arcsinh",
  cofactor = 5
)
}
\arguments{
\item{input.file}{(Required) Path to input file}

\item{channels}{Default: NULL. Specify to subset data. Please provide a vector with full channel name (e.g. c('142Nd_CD19', '154Sm_CD3')). Channel names can be found using the print_parameters function.}

\item{transform}{'arcsinh' or 'none'. 'arcsinh' performs an arcsinh transformation. 'none' performs no data trasformation}

\item{cofactor}{Cofactor for arcsinh transformation (Default: 5)}
}
\description{
This function reads an FCS file and returns a data frame of read counts
}
\examples{
data <- load_data('example.fcs')
}
