% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expressions.R
\name{load_multi_samples}
\alias{load_multi_samples}
\title{Loading BinaryClust results of multiple datasets}
\usage{
load_multi_samples(file.list = NULL)
}
\arguments{
\item{file.list}{(Required) A list of files in comma-separated format. Please see tutorial for details.}
}
\description{
This function concatenates multiple results of cluster_summary for differential analysis. Returns a data frame.
}
\examples{
multi.data <- load_multi_samples(file.list)
}
