% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_binary_abundances}
\alias{plot_binary_abundances}
\title{Plotting abundances based on binary classification results}
\usage{
plot_binary_abundances(
  binary.summary = NULL,
  remove.unclass = FALSE,
  xlabel = "",
  ylabel = "Percentage Abundance",
  title = "Cell Abundances"
)
}
\arguments{
\item{binary.summary}{(Required) A data frame containing abundance results. Expects results from binary_summary}

\item{remove.unclass}{Option to remove unclassified cells from plot. (Default: FALSE)}

\item{xlabel}{Label for x-axis (Default: NULL)}

\item{ylabel}{Label for y-axis (Default: 'Percentage Abundance')}

\item{title}{Plot title (Default: 'Cell Abundances')}
}
\description{
This function plots the abundanceresults of binary classification. Returns a graph.
}
\examples{
plot_binary_abundances(binary.summary)
}
