% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_binary_median}
\alias{plot_binary_median}
\title{Plotting median expressions based on binary classification results}
\usage{
plot_binary_median(
  binary.summary = NULL,
  remove.unclass = FALSE,
  title = "Median Expressions"
)
}
\arguments{
\item{binary.summary}{(Required) A data frame containing abundance results. Expects results from binary_summary}

\item{remove.unclass}{Option to remove unclassified cells from plot. (Default: FALSE)}

\item{title}{Plot title (Default: 'Median Expressions')}
}
\description{
This function plots the median expressions of all markers after binary classification. Returns a graph.
}
\examples{
plot_binary_abundances(binary.summary)
}
