% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cluster_heatmap}
\alias{plot_cluster_heatmap}
\title{Plotting heatmaps for post-binary classification clustering results}
\usage{
plot_cluster_heatmap(data = NULL, cluster.results = NULL, cell.type = NULL)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{cluster.results}{(Required) Clustering results in data frame format. Expects results from cluster_subsets}

\item{cell.type}{(Required) Cell subset to be plotted. This must be an entry from the binary classification specification}
}
\description{
This function plots a heatmap of clusters of a user-specified cell type. To be used after running cluster_subsets. Returns a graph.
}
\examples{
plot_cluster_heatmap(data, cluster.results, 'T Cells, CD4')
}
