% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expressions.R
\name{plot_diff_abundances}
\alias{plot_diff_abundances}
\title{Plotting abundances of multiple datasets}
\usage{
plot_diff_abundances(
  multi.data = NULL,
  column = "Cell.Type",
  remove.unclass = FALSE,
  style = "bar"
)
}
\arguments{
\item{multi.data}{(Required) A data frame of clustering results. Expects results from load_multi_samples}

\item{column}{Choose between plotting Cell.Type or Cell.Subtype (Default: Cell.Type)}

\item{remove.unclass}{Option to remove unclassified cells from plot (Default: FALSE)}

\item{style}{Choose between two styles of plots, bar or dot (Default: bar)}
}
\description{
This function plots the abundance of multiple datasets. Returns a graph.
}
\examples{
plot_diff_abundances(multi.data)
plot_diff_abundances(multi.data, column = 'Cell.Subtype', style = 'dot')
}
