% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expressions.R
\name{plot_diff_exp}
\alias{plot_diff_exp}
\title{Plotting differential analysis results}
\usage{
plot_diff_exp(diff.results = NULL, plot_p_values = TRUE)
}
\arguments{
\item{diff.results}{(Required) A list of data frames from differential analysis. Expects results from calculate_diff}

\item{plot_p_values}{Option to display significance values on the heatmap (Default: TRUE)}
}
\description{
This function plots the results of differential analysis and returns a summary heatmap
}
\examples{
plot_diff_exp(diff.results)
}
