% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_NRS}
\alias{plot_NRS}
\title{Calculating and plotting non-redundancy scores (NRS)}
\usage{
plot_NRS(data = NULL)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}
}
\description{
This function calculates and plots non-redundancy scores of all the markers in data. Returns a graph.
}
\examples{
plot_NRS(data)
}
