% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_TSNE}
\alias{plot_TSNE}
\title{Plotting binary classification results on t-SNE coordinates}
\usage{
plot_TSNE(
  tsne.results = NULL,
  binary.results = NULL,
  column = "Cell.Type",
  title = "",
  facet = FALSE
)
}
\arguments{
\item{tsne.results}{(Required) t-SNE object. Expects results from run_TSNE}

\item{binary.results}{(Required) Results from binary classification in data frame format. Expects results from binary_class}

\item{column}{Choose between plotting Cell.Type or Cell.Subtype (Default: Cell.Type) Cell.Subtype only availble for updated clustering results}

\item{title}{Title of the plot (Default: NULL)}

\item{facet}{Breaks plot into facets (Default: NULL)}
}
\description{
This function plots t-SNE data coloured by binary classification results. Returns a graph.
}
\examples{
plot_TSNE(tsne.results, binary.results)
plot_TSNE(tsne.results, updated.cluster.results, column = 'Cell.Subtype', facet = TRUE)
}
