% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_TSNE_marker}
\alias{plot_TSNE_marker}
\title{Colouring t-SNE results with marker expressions}
\usage{
plot_TSNE_marker(
  tsne.results = NULL,
  data = NULL,
  marker = NULL,
  lower = NULL,
  upper = NULL,
  palette = "Purples",
  title = ""
)
}
\arguments{
\item{tsne.results}{(Required) t-SNE object. Expects results from run_TSNE}

\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{marker}{(Required) Marker to be coloured}

\item{lower}{Lower bound of colour spectrum (Default: Lowest value of marker expression)}

\item{upper}{Upper bound of colour spectrum (Default: Highest value of marker expression)}

\item{palette}{ggplot2 colour spectrum (Default: 'Purples')}

\item{title}{Title of plot (Default: NULL)}
}
\description{
This function colours t-SNE results by marker expressions. Returns a graph.
}
\examples{
plot_TSNE_marker(tsne.results, data, marker = 'CD45')
}
