% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_UMAP}
\alias{plot_UMAP}
\title{Plotting binary classification results on UMAP coordinates}
\usage{
plot_UMAP(
  umap.results = NULL,
  binary.results = NULL,
  column = "Cell.Type",
  title = "",
  facet = FALSE
)
}
\arguments{
\item{umap.results}{(Required) UMAP object. Expects results from run_UMAP}

\item{binary.results}{(Required) Results from binary classification in data frame format. Expects results from binary_class}

\item{column}{Choose between plotting Cell.Type or Cell.Subtype (Default: Cell.Type) Cell.Subtype only availble for updated clustering results}

\item{title}{Title of the plot (Default: NULL)}

\item{facet}{Breaks plot into facets (Default: NULL)}
}
\description{
This function plots UMAP data coloured by binary classification results. Returns a graph.
}
\examples{
plot_UMAP(umap.results, binary.results)
plot_UMAP(umap.results, updated.cluster.results, column = 'Cell.Subtype', facet = TRUE)
}
