% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{print_parameters}
\alias{print_parameters}
\title{Printing the parameters of an FCS file}
\usage{
print_parameters(input.file = NULL)
}
\arguments{
\item{input.file}{(Required) Path to input file}
}
\description{
This function reads an FCS file and prints out the parameter page of the flowCore flowFrame object. It is intended to aid feature selection.
}
\examples{
print_parameters('example.fcs')
}
