% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{run_BinaryClust}
\alias{run_BinaryClust}
\title{Running BinaryClust with default settings}
\usage{
run_BinaryClust(input.file = NULL, class.file = NULL)
}
\arguments{
\item{input.file}{(Required) Path to input file}

\item{class.file}{(Required) Path to classification file in comma-separated format. Please see tutorial for details.}
}
\description{
This function performs all major steps of BinaryClust using default settings. Returns a list of data frames.
}
\examples{
BinaryClust.results <- run_BinaryClust('example.fcs', class.file = 'cell_types.csv')
}
