% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{run_TSNE}
\alias{run_TSNE}
\title{Calculating t-SNE coordinates for data}
\usage{
run_TSNE(
  data = NULL,
  scale = TRUE,
  center = FALSE,
  perplexity = 30,
  n_iter = 1000
)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{scale}{Default: TRUE. Option to standardise data prior to t-SNE calculations}

\item{center}{Default: FALSE. Option to center data prior to PCA within the Rtsne algorithm}

\item{perplexity}{Value for perplexity (Default: 30)}

\item{n_iter}{Number of iterations (Default: 1000)}
}
\description{
This function calculates t-SNE coordinates on data. Random seed is set to 1. Returns full t-SNE object.
}
\examples{
tsne.results <- run_TSNE(data)

t-SNE coordinates can be extracted under Y:
tsne.coordinates <- tsne.results$Y
}
