% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{run_UMAP}
\alias{run_UMAP}
\title{Calculating UMAP coordinates for data}
\usage{
run_UMAP(data = NULL, scale = TRUE)
}
\arguments{
\item{data}{(Required) Data in data frame format. Expects results from load_data}

\item{scale}{Default: TRUE. Option to standardise data prior to UMAP calculations}
}
\description{
This function calculates UMAP coordinates on data. Random seed is set to 1. Returns full UMAP object.
}
\examples{
umap.results <- run_UMAP(data)

UMAP coordinates can be extracted under layout:
umap.coordinates <- umap.results$layout
}
