% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{update_cluster_results}
\alias{update_cluster_results}
\title{Updating clustering results with filled subtype column}
\usage{
update_cluster_results(cluster.results = NULL, completed.summary = NULL)
}
\arguments{
\item{cluster.results}{(Required) Cell type results in data frame format. Expects results from cluster_subsets}

\item{completed.summary}{(Required) Cluster summary in data frame format. Expects results from cluster_summary with a completed Cell.Subtype column}
}
\description{
This function takes in the filled cell subtype column of a clustering summary and updates the associated all-cell clustering results. This function is intended for the UMAP plotting function. Returns a data frame.
}
\examples{
updated.cluster.results <- update_cluster_results(cluster.results, completed.summary)
}
